Fixed centering bug, optimized DOM references, changed naming format, added demo
.width() and .height() were being called on elements of the 'center*'
class but not necessary the correct ones. Not an issue if you only
center one element, but breaks if more than one element use the
centerVertical or centerHorizontal  classes.

Class names are defined at the beginning to be more DRY and to run
faster, as the DOM only has to be traversed once now.

'midway' is prepended to the class names to make them domain specific.
'center' is removed because it's little redundant when 'midway' is in
the name. The change from camel case to dashes was to make it match CSS
notation.

I also changed the onload() call format. I think this looks cleaner